Restructure of new recipe form
Decided to restructure the new recipe form. Will have the user enter the recipe information, and then on a new screen have them enter a name for the ingredient group along with the ingredients.
With the new recipe formated correctly, next is to get it to save and open the screen for the user to enter the ingredient group and the ingredients.